Split SvgCubicCurveSegment
