+++ Source/Paths/SvgCubicCurveSegment.cs
@@ -1,11 +1,8 @@
 using System.Drawing;
-#if !NO_SDC
-using System.Drawing.Drawing2D;
-#endif
 
 namespace Svg.Pathing
 {
-    public sealed class SvgCubicCurveSegment : SvgPathSegment
+    public sealed partial class SvgCubicCurveSegment : SvgPathSegment
     {
         public PointF FirstControlPoint { get; set; }
         public PointF SecondControlPoint { get; set; }
@@ -21,36 +18,6 @@ namespace Svg.Pathing
             : this(isRelative, NaN, secondControlPoint, end)
         {
         }
-
-#if !NO_SDC
-        public override PointF AddToPath(GraphicsPath graphicsPath, PointF start, SvgPathSegmentList parent)
-        {
-            var firstControlPoint = FirstControlPoint;
-            if (float.IsNaN(firstControlPoint.X) || float.IsNaN(firstControlPoint.Y))
-            {
-                var prev = parent.IndexOf(this) - 1;
-                if (prev >= 0 && parent[prev] is SvgCubicCurveSegment)
-                {
-                    var prevSecondControlPoint = graphicsPath.PathPoints[graphicsPath.PointCount - 2];
-                    firstControlPoint = Reflect(prevSecondControlPoint, start);
-                }
-                else
-                    firstControlPoint = start;
-            }
-            else
-                firstControlPoint = ToAbsolute(firstControlPoint, IsRelative, start);
-
-            var end = ToAbsolute(End, IsRelative, start);
-            graphicsPath.AddBezier(start, firstControlPoint, ToAbsolute(SecondControlPoint, IsRelative, start), end);
-            return end;
-        }
-
-        [System.Obsolete("Use new AddToPath.")]
-        public override void AddToPath(GraphicsPath graphicsPath)
-        {
-            AddToPath(graphicsPath, Start, null);
-        }
-#endif
 
         public override string ToString()
         {
